feat: Integrate Mentions with MessageContent
New `Mention`s can be now sent from `ConversationsUiService.addTextMessage`.
They are added to `MessageContent`s of `MessageData` and adjusted for the specified charset.
We differentiate between two charsets: UTF-16 and default. When we want to send a message,
we specify that the mentions in that message should be adjusted for UTF-16. When we receive
one, we need to re-adjust the mentions for our default charset (UTF-8 on Android). Sending
and receiving messages use common fuctionality in MessageData: `messageContent`. Before this
method was used the same way no matter the direction. Now we have to specify which way
the message will go after adjusting. This creates a special case for editing messages.
An edited message is both re-displayed on the our side, as if we received it, and it is sent
to other devices. In this commit we handle only the sending side. The re-displaying of
an edited message needs additional work.

I added a few unit tests, but usefulness of some of them is a bit limited: the default charset
for Java and Scala is UTF-16. Only when we're actually on Android, the default charset becomes UTF-8.