fix: check for potentially out of sync push token
Due to a race where we might accidentally unregister our own push
token from the BE without noticing, it can be that some users in the
wild have no push token.

This commit implements functionality to fetch all registered push tokens
on the BE for a given user, upon which the PushTokenService will compare
these tokens with the one it thinks is registered. If there is no match,
then it will re-register the current device token with the backend.

This commit only provides the above described functionality, but does
not schedule the check. This will be determined in the Android (UI)
project - the best time to do it would be on app start and app upgrade.

partly fixes: https://github.com/wireapp/android-project/issues/263