fix: add job for checking registered PushToken
This commit provides an Evernote Job for, and executes, the check to see
if the current user has a correctly registered push token.

The evernote job handles scheduling, back-off and rescheduling of the
work, while the job itself will be triggered from inside the
WebSocketBroadcastReceiver (which I will now rename to
StartAndUpgradeBroadcastReceiver). This should be sufficient to
ensure that anyone with the bug will be re-registered the next time
they start the app after upgrading, and every device restart should
be a regular enough check to make sure the token is still valid.

partly fixes: https://github.com/wireapp/android-project/issues/263