refactor: move sync failure logic to handlers
Since we handle side-effects for successful syncs inside the sync handlers,
it also makes sense that we handle the side-effects for failures too. The
only problem then is that the sync handlers need information about what
attempt number this request is, and when the request was supposed to be
started, so that the handler can decide if it needs to give up or timeout.

This commit passes makes this information available for the sync handlers,
and ensures that the sync executor/service fails in only very generic ways.