fix: use Signal#publish properly
It turns out that the currentContext parameter in the Signal#publish
method is exactly that, a hint to what the currentContext is, in which
publish is called. This is then used to avoid posting a future to the
Signal's subscription later on down the chain so that the subscription
can then be executed synchronously.

Previously, we thought that the publish method was there to specify an
execution context on which the subscriptions should be executed. This
assumption has proven to be false and this PR fixes the areas where these
false assumptions where made.

Luckly there were no serious side-effects to this mistake, and hence no
ticket to link here...