chore: introduce public logging to log-in flow
This commit continues some of the work to enable safe logs in the public
version of the app. In particular, it focuses on any code that may be
invoked within the log in flow and initial loading of the contacts list.

Some very verbose print statements were also removed in this commit to
reduce the amount of logging on log in