feat: Add storage and support for read receipts (#453)
Read receipts are also Confirmation messages like Delivery receipts but with the flag set to READ.
Delivery receipts were also renamed from Receipt to DeliveryReceipt to avoid ambiguity.
A dedicated storage for these receipts was added since read receipts, unlike delivery receipts, will be used in group conversations so a list of users who read a certain message should be stored along with the timestamp of the event.
Backend has a properties endpoint which allows storage of json objects given a key. We are using this to store the setting for the read receipts.
Also, an event each time a property gets changed is triggered which we also now process in PropertiesService.