feat: Replace id extractors of classes held in DB  with compile-time information about their ids. (#461)
When accessing storage we often use "id extractors" (or "key extractors") to get the id of an object we're manipulating.
The id extractor accesses the db and retrieves the id from there. This is necessary because until now at this point in code we knew nothing about the type of the object. It was a generic type `V`, associated with a `Dao` class and only from there we were able to get any information about it, at run-time.
But if we change the signature of case classes of objects we held in the db, we can easily access their ids at compile-time, rendering id extractors unnecessary. It should lower the amount of times we access storage.