feat: Enable custom icons. (#1960)
If `custom/custom.json` exists, it is used as a config file instead of `default.json` and it can specify which icons will be replaced
by custom ones, taken from a specified Git repository. The repository is cloned and the original icons are replaced by their customized
versions (in the `src` directory, so there is no way to get the original ones back). The user has to provide icons for all resolutions
(ldpi, mdpi, hdpi, ...) and only a hardcoded subset of icons can be replaced.