fix: Catch ArrayIndexOutOfBoundsException on Android 7.0 (#1974)
`ViewGroup.measureChildWithMargins` sometimes throws an exception on Android 7.0 (SDK 24). It doesn't happen on any other Android version,
and from Google Play Console it seems that even for Android 7.0 it happens only on some devices. But the amount of reports is significant,
so after investigation I decided to simply wrap that method in try/catch on Android 7.0. If we ever be able to reproduce that bug, or get
more information, I will get back to it.

I also rewrote the `onMeasure` method (which calls `measureChildWithMargins`), so it does not perform some calculations twice.