Fix: CreateConversationSettings UI elements not reacting to events (#2000)
Many UI elements in the "create conversation" screen were not responsive to user event after switching to another fragment and coming back.

In particular, the "show/hide" arrow for settings, and the list of users (it would not update after switching guests on/off). This was caused by calling `viewHolder.foreach` in the body of a `returning(... viewholder ...)`. With this code, the foreach would be called only once when the fragment is created. If the view is then destroyed and re-created, the new view inside the viewholder would not have been "visited" by the foreach.

The fix is to call the `viewholder.foreach` in the `onCreateView` method, which is called again every time the view is re-created (and the new individual views are already available in the view holder).