feat: Add custom fields to the conversation participant's details
The conversation participant's details screen is changed into a scrollable view which can display
a list of custom fields in the space between the chathead and the bottom bar.
The bottom bar was simplified: It's 'floating', meaning that it's always visible, no matter
the tab and scrolling. There is always the "..." icon on the right side, even if there's only
one option to choose from the menu.
There are still some design details to fix: fonts, distances, the "Remove Member" icon.