fix: A partner should be treated as a guest when displaying the bottom bar (#2020)
Fixes a bug when the user (who is a partner, not a guest) opens group conversations' details and from there
goes to a single participant details. The only possible action for a partner on this screen should be
"Open Conversation" which appers on the left button. The right button ("...") should be invisible.
Because of the bug it was visible, but when clicked, no options appeared.