fix: Open the Okta page directly from the Welcome page (#1998)
Before, we moved to the Registration page first, so when the user clicked "Go back" on the Okta page
they were taken back to registration. By design, it should be the Welcome page.
Also, now "Go back" erases the token, preventing a corner case when the user clicks for the second time
on the link, while being on the OKta page already.