fix: Opening the participant's details from the device confirmation dialog should land on the Devices tab (#2040)
fixes AN-6094

If a conversation is degraded, a dialog appears and one of the options is to go to the other participant
details to see the unconfirmed device. In that case, the participant's details should automatically
open the Devices tab, but it was opening the Details tab instead. That was fixed.

Also, the dialog didn't always display the other participant's name, the sorting of users' names
in system messages was broken, and the shield displayed in a system message was too wide.
I fixed that as well.