fix: SendConnectRequestFragment should show "..." instead of "-" as the right footer menu action
`SendConnectRequestFragment` is used both for sending requests and as a screen the user when they open
an unconnected chat group member details. In the second case, the right button on the footer menu is visible.
Until now it was the "-" icon, indicating the possibility to remove the user from the group.
Now, because of changes of how we display the footer menu, this is changed to the universal "..." icon.
Clicking on it opens the bottom popup menu and from there the user may remove the other person from the group.