fix: Reopening the SSO Login from an SSO link should reset the page
Until now bringing the app from the background tothe front simply displayed the last screen the user was on.
With the introduction of deep links this will have to change. Just as with opening the app from a notification,
the app will have to figure out what screen should be displayed.
If the user clicks on an SSO link while the app is in the background and already on the SSOWebViewFragment,
the fragment should not just be displayed, but it should be reloaded.