fix: The SSO dialogs should not pile on each other
If the user went back and forth between the Welcome Page and subsequent login/registration pages,
the SSO dialog was sometimes created twice for every page, but when the user clicked `cancel`,
only one was really dismissed. The other didn't appear, but it was also not gone: when the user
went back and forth again and saw the dialog again, now two clicks on `cancel` were required
to dismiss it. Then three, then four, etc.
I fixed the creation of the second dialog, and also changed the way listeners are added and
removed from the dialog (they should be handled in `onStart` and `onStop`).