Fixes to reopening the SSO Login page and dialog (#2046)
* fix: Reopening the SSO Login from an SSO link should reset the page

Until now bringing the app from the background tothe front simply displayed the last screen the user was on.
With the introduction of deep links this will have to change. Just as with opening the app from a notification,
the app will have to figure out what screen should be displayed.
If the user clicks on an SSO link while the app is in the background and already on the SSOWebViewFragment,
the fragment should not just be displayed, but it should be reloaded.

* fix: The SSO dialogs should not pile on each other

If the user went back and forth between the Welcome Page and subsequent login/registration pages,
the SSO dialog was sometimes created twice for every page, but when the user clicked `cancel`,
only one was really dismissed. The other didn't appear, but it was also not gone: when the user
went back and forth again and saw the dialog again, now two clicks on `cancel` were required
to dismiss it. Then three, then four, etc.
I fixed the creation of the second dialog, and also changed the way listeners are added and
removed from the dialog (they should be handled in `onStart` and `onStop`).