fix: Properly initialize the app before handling conversation and profile deep links
Currently deep links work only if the app was previous opened and now works in the background.
This is because we handle all deep links in `MainActivity`. If the app is not initialized already,
we try to switch the conversation or open a profile before `MainPhoneFragment` is set up.
This PR delays handling of these two types of deep links, moving them to the `onStart` method
of `MainPhoneFragment`.