DeepLinks. (#2051)
Deep links for switching conversations and opening user profiles

Now clicking on a deep link from outside Wire can open the app and lead the user to a conversation or a user profile. The corner cases are also handled. Trying to open a conversation with an unconnected user results in an error message. Opening a user profile of an unconnected user leads to the Send Connect Request screen.