fix: The conversation list does not wait for syncing to finish
This is a partial fix to AN-6149 where the app is stuck on syncing. When something goes wrong during the sync, we should log the error
and finish (failed jobs might be rescheduled). The conversation list ait for that and only then the conversations are displayed.
But it is possible that the sync takes longer than usual or - as in the bug case - that it's stuck. Then the conversation list
does not appear at all. From the user's perspective it's better for the conversations to appear and the sync should run in the background.

We still don't know what exactly went wrong during the sync, but I think this part of the fix could be applied anyway.