feature: use `LOCAL_ZMESSAGING_VERSION` env for local builds
Building sync engine with sbt creates a file named `141.x.x-SNAPSHOT`. The name then needs to be hardcoded in the `build.gradle` file, and there is a risk that it will be committed by mistake.

This commit modified the `build.gradle` so that a local variable can be set to use a snapshot version of sync engine.