feat: Status notifications
The status (availability) affects the notifications now. If the status is set to 'Away'
the notifications are not displayed at all. If it's 'Busy' the notifications are displayed
only if certain requirements are met. Calls also are affected in a similar way.

I also made a small change to `app.gradle`, because otherwise I was not able to use SE.

SE part: https://github.com/wireapp/wire-android-sync-engine/pull/538