fix: Fixes an inconsistency in how a change of the Conversation Muted status is reported between Android and other platforms (#547)
The number sent from Android - which identifies the new conversation's muted state - was different from other platforms and couldn't be interpreted
properly on them. Also, the `muteTime` property was set to a time in the past, which on iOS was interpreted as that the event has already taken
place and was ignored.

fixes https://wearezeta.atlassian.net/browse/AN-6224
fixes https://wearezeta.atlassian.net/browse/AN-6210