fix: Ensure that the app is initialized or otherwise don't use its modules. (#2143)
Apart form the main path, whne the user opens the app manually, the app can be initialized via intents,
handled by services and receivers, such as `FCMHandlerService` and `OnBootAndUpdateBroadcastReceiver`.
The app should be initialized form them before any action can be taken as the result of receiving
the intent. However, sometime the initialization is not possible and in such cases the receiver should
log a message and close gracefully, without crashing.