Fix: verification code is send if password is valid (#2122)
* Verification code is send if password is valid

I realized that when creating an account, if I type a invalid password (based on the policy) I would still receiving an email with the verification code. And if I keep making the same mistakes, I would continue receiving more emails until a correct password was typed. I correct this for making the validation at first

* Implementation of requested changes