fix: Chatheads in participants lists display placeholders instead of avatars
With new assets introduced, `ChatHeadView` was rewritten. The old version cleared the displayed image
every time it was asked to load user data, because it might happened that the avatar was updated.
In the new version, however, it only made the avatar disappear and be replaced by a placeholder.

The new version should handle avatar updates differently, but it doesn't. I will work on that in
another PR.

I also made some slight refactoring: A few placeholders were created for each avatar, even though
only one was necessary.