fix: A timer appears over a chathead of another participant
A timer icon over the avatar of another user indicates that we sent a connect request to that person, or that they sent a connect request to us.
Instead, the timer icon appeared also on avatars of all participants in the given conversation who were not connected to us.
That was fixed.