fix: WorkManagerSyncRequestService NPE when used too early (#2185)
This is a fix for a NullPointerException reported by the Google Play Console.
`WorkManagerSyncRequestService.doWork()` method is called before the app is initialized and when the method tries to inject the SyncHandler and then use it, the SyncHandler is still not initialized.
I wasn't able to reproduce this bug, but the behavior looks very similar to what I worked on recently
in `FCMHandlerService` and `WebSocketService`. I added a call to `WireApplication.ensureInitialized()` before the SyncHandler is injected.