fix: It shouldn't be possible to create multiple group conversations by clicking the button many times quickly
It is technically possible to create many conversations with the same participants: fake 1-to-1, and real group convs.
This possibilty leads to a bug when the user clicks the "Create conversation" button many times quick enough that
if the app experiences a delay it is possible to create many conversations at once. To prevent that, clicks on all
the buttons used for creating conversation - in the "Create group" UI, "Create guestroom", and fake 1-to-1
conversations with other team members, are now ignored while the creation of the ocnversation is not finished
after the first click.