fix: Asking for audio permissions on the quick recording leaves the app in an invalid state (#2202)
The quick recording means that the user click the recording icon and holds the finger down. The audio is
recorded to the moment the finger is lifted, and after that it can be dismissed or sent.
However, if we need to ask for permissions first, before the start the recording, we display a dialog window
and, of course, that means that the user has to lift the finger from the recording icon and click
"Ok" or "Cancel". But if "Ok" is clicked that leaves us in a situation when the recording is started,
but the finger is already lifted.

I moved the query for permissions - and the dialog window - one layer up, to `ConversationFragment`.
Now, if we need to ask for permissions, then even if they are granted, we cancel the recording.
The user has to click the recording icon again and this time the recording will start without the window
popping up.

fixes: https://wearezeta.atlassian.net/browse/AN-6288