fix: Quick recorder exception (#2207)
* fix: Handle the exception thrown by MediaRecorder

Recording can sometimes fail, for many reasons, and we learn about it only when we stop it.
In that case we should delete the file and abort. So I made a fix where I handle this case.
It means that on some devices it won't be possible to use quick recording, but at least there
will be no unhandled exception which crashes the view.
Full recording still works, though. It uses the recorder in a different way.

Look here for a short discussion:
https://stackoverflow.com/questions/10147563/android-mediarecorder-stop-failed

* fix: Make quick recording use the same functionality as the standard recording

Standard recording was rewritten in Kotlin recently. Quick recording is in Scala, and it was failing on some of the devices.
This fix makes the quick recording use the same Kotlin code as the standard recording.
It also provides the code which connects Scala and Kotlin for future use.