fix: The guests' banner updated when the conversation view is invalid
The guests' banner shows up if we have guests in the conversation. Every time the conversation view is displayed,
we check if we should display or hide the banner. The code which informs us that we should do it is in Java and
it sometimes is called even if the conversation view is not actually on top - the participants list may cover it.
In that case, updating the banner causes an exception. It doesn't crash the app, but it causes a delay which
fails the tests.

I added an additional check to the old code, ensuring that the banner is updated only if the conversation view
is on top.