feat: Bumping the max number of conversation participants to 500 (#2204)
* feat: Bumbing the max number of conversatino participants to 500

* fix: The guests' banner updated when the conversation view is invalid

The guests' banner shows up if we have guests in the conversation. Every time the conversation view is displayed, we check if we should display or hide the banner. The code which informs us that we should do it is in Java and it sometimes is called even if the conversation view is not actually on top - the participants list may cover it.
In that case, updating the banner causes an exception. It doesn't crash the app, but it causes a delay which fails the tests.

I added an additional check to the old code, ensuring that the banner is updated only if the conversation view is on top.

* feat: Remove pseudo-random creation of a conversation's avatar

I also changed the strings informing about the maximum number of participants, so that now they use `ConversationController.MaxParticipants`.