fix: Record and send video files.
Until now the process of reconding and sending videos looked as following:
1. Create a path to the video file in our media cache directory.
2. Provide that path to the intent sent to the camera.
3. The camera records the video to the file with the path we specified.
4. We use the specified file to send the video.

On some devices the access to the provided location was not possible. The bug didn't appear to be connected to the Android version,
and on affected devices it didn't matter what permissions were given to the app or the cache directory.
But also, our way of recording videos was different than usual. The usual way is to let the camera create a file at its own
location, and then use it. We chose another approach because of security reasons - we wanted to keep our videos in our
internal directories.

The fix lets the camera write to its own file, and then copies the file to our cache directory and deletes the original one.