fix: A conversation badge tries to display a non-existent glyph
A conversation badge can be either a text, a glyph, or a style kit view.
Only one of these should be displayed. But now, initially, we load all three,
and only then hide two of them. In a manual test on the dev build it leads
to a crash which I wasn't able to reproduce on Internal. But even if that's
some sort of a glitch, we display a lot of conversation badges and it's an
easy and useful optimization to just display the correct type and to never
load the other ones.