fix: Decrease message data loaded to memory when opening/browsing a conversation (#2229)
Currently opening a conversation results in a query to the database which can load up to 2MB of data
at once in the memory. We suspect we load too much which may trigger garbage collection and random
exceptions connected to the lack of memory. In such cases the user may see a blank page or even
contents of the old conversation.

The fix has two steps:
1. The size of the data loaded at once is decreased.
2. If an exception is thrown, we try to recover with an even smaller paged list.

It is still technically possible that we won't be able to liad the data even with the smaller
page list. It might mean that the device is too wak to handle the application.