Change meaning of return value when running security checklist
Before, the boolean returned from `run()` had the meaning that all checks passed. This method has been renamed to `canProceed` to indicate that app should block or not. This solves the issue where the blocking dialog would only block as long as its context was alive. If the context disappeared, such as when the activity finishes, then it wouldn't be blocking anymore and lead to some other issues.