fix: Invalid system message about adding someone to the conversation (#2220)
fixes: https://wearezeta.atlassian.net/browse/AN-6299

Sometimes when the user creates a new conversation, a MEMBER_JOIN system message is received:
"You added X to the conversation" where X is the name of the same user.
I wasn't able to reproduce that situation, but I found a place in `MemberChangePartView`
which may lead to that invalid text being displayed. The creation of a conversation
may be treated as if the creator added him/herself to it, and then a system message may be
created which has the selfId both as the sender, and the other user id.

I believe this system message is redundant and can simply be ignored.
The right text should be displayed in `ConversationStartPartView` instead.