fix: Image rotation
Some images have their rotation written in the Exif information, e.g. they may appear landscape,
width longer than height, but they should be displayed as a portrait. While sending image
as an asset in a conversation we lose that information.
Trying to maintain the Exif info (and then use it on the receiver side) would affect too much
code, so we chose an alternative approach: Before sending, we first check the Exif info. If the
image needs rotation, we make a rotated copy of it - which no longer requires Exif - and send
the copy. It makes the sending process slower, but only in a corner case.