fix: Photos from the internal camera are sometimes failing to be sent (#2255)
When sending an asset we need to copy the data from its source to the cache in the app,
and then we use the file from the cache in the rest of the process. Sometimes, fo example
when taking a photo directly from within the app, the source points already to the file
in the cache. Currently, in such situation, we still copy it (it should be unnecessary,
but this is a bigger issue than this bug). Depending on the device copying data when
the source and the target point to the same place, may result in an exception.

I fixed it by first reading the whole data in, closing the input stream, and only then
writing the data to the target.