fix: Delay loading the current conversation for after the sync
When Wire is stopped and opened again, the initialization of the UI interferes with syncing of new events,
e.g. the fist conversation contents are loaded, even though it's not immediately displayed to the user.
Making the initialization lighter is a big issue, but one thing that can be done easy is to delay that
loading of the first conversation to after the syncing is done.