fix: Save image to gallery (#2260)
When saving a photo to the gallery (or any assets to Downloads) we sue the asset's id to create the filename.
Apparently, sometimes asset id may contain "/". It's probably connected to the way UUID are generated by Android,
because it happens only a small number of devices.
This fix looks for "/" in the generated filename and replaces it with "_". The asset id stays untouched.