fix: Prevent the password screen from being displayed when the password was already provided
When the app asks for the password, the control is given to Android - the app's MainActivity
is stopped, and then started again when the control goes back to the app. On `onStart`
we perform the check if the password is correct. But due to the unsure order of methods
being called when leaving and entering the app, it was possible to check for the password
validity *before* the information about it was received from Android. In that case,
the app displayed the password screen again.

Because of the same issue when the user displayed their profile the app treated that as if
it went into background and asked for the password when the user went back to the conversations
list.