fix: Make sure the AppLockActivity is not created from LaunchActivity
In order to ask for the Android password/PIN we have to start `AppLockActivity`. To do that, we need to
provide a reference to the parent activity. Right now we do this by choosing whichever activity triggered
the request to ask for the password/PIN. At the start of the application, that's `LaunchActivity`.
`LaunchActivity` triggers `AppLockActivity`, but it also chooses to start `MainActivity` or `AppEntryActivity`,
and then closes itself.
Depending on the device, the order in which these activities are started may differ. It may lead to the
situation when Main- or AppEntryActivity goes on top of `AppLockActivity` and then freezes, because it's
waiting for user authentication, which never comes, because `AppEntryActivity` is not on top.

The fix simply ignores `LaunchActivity`. We know for sure that just after `LaunchActivity` another activity
will be started, and we can start `AppLockActivity` from that one, ensuring that it will always be on top.