fix: Ephemeral videos and audios were never deleted
In theory, ephemeral video and audio recordings need to be treated a bit differently from other assets:
on the receiver's side they should be treated as "read" (and so, the expiration clock should start on them)
only when they are played.
The old solution to it was quite complex it broke down after we switched to new assets. The bug was that
the videos and audios were never treated as "read", so the clock never started.

This fix makes videos and audios work like any other asset. The delayed expiration issue will be addressed
in another PR.

I also removed some unused code.