Decrease test iterations from 42k (!) to 1000
This test was constantly failing on CI. I'm not sure what this test is trying to test, my guess is that it tests that "nothing crashes with a high number of events".

I think it probably hits some memory limit by creating so many objects, GC kicks in and slows everything down to a halt, making it time out afte 15s.