feat: Conversation folders (#2320)
* Data model

* Basic FoldersService and storages

* Separate being in a folder from being a favourite

Conversations can be in multiple custom folders. Adding or removing a conversation from a custom folder does not remove it from favorites. Adding or removing a conversation from favourites it does not remove it from custom folders.

* Rename folders method and add `moveToFolder`

* Use the name "customFolder" for methods

* A signal to present folders to UI and fixed unit tests

* Feat: Folders functionality added to ConversationListController (#2325)

Folders functionality added to ConversationListController,
changes requested in the review and the rewritten signal

* more changes requested in the review

* methods for adding/removing folders and conversations to/from folders