fix: Bottom navigation bar highlights
The main problem was that the bottom navigation bar was updated in the background thread - it sohuld be done in UI.
I changed that, but I also moved some code apart so that we to the background, because the other error showing in
the logs was that we calculate in UI two heavy signals: one for checking if we should display
the 'archive' icon, and the other to check if we have regular conversations to display. I merged them into one
and moved to `ConversationListController` which works on the background thread.