fix: If obfuscation the "Recents" list is impossible, remove Wire from the list
Implements an alternative solution to obfuscation of the "Recents" list for Android 5.1 where the obfuscation doesn't work:
Instead the app is completely removed from the list of "Recents" apps. It can be still accessed simply by clicking the icon
on the main screen of the device.

I rewrote the old `SecurityLifecycleCallback` into `ActivityLifecycle`. Now it exposes a signal and classes which want to
react to going to the background may subscribe to the signal (that's the process reversal - `SecurityLifecycleCallback`
called `SecurityPolicyChecker` directly). In result, `ActivityLifecycle` serves now a similar role to `UiLifeCycle`, but
it doesn't rely on `BaseActivity` calling `acquireUi` and `releaseUi`. I think in another PR we should replace `UiLifeCycle`
with it.