feat: Wipe data after failed N attempts at typing in the password
This code does not count failed attempts, because the app is not getting notified about them for some reason,
but the number of times the user cancelled typing in the password. With this in use, the user has a choice:
either to stay on the Android's password page and typing in the right password (the policy uses the
exponential backoff, so with many failed attempts this may mean being locked out for a long time) or if
they hit "Cancel" enough times, the Wire data will be wiped.