refactor: A migration to AndroidX in preparation for using BiometricPrompt
This commit is huge and quite messy. There are three things going on here at once:

1. The Android security policy functionality is broken or at least doesn't work as the documentation claims.
I moved setting the policy from `SecurityPolicyChecker` to `BaseActivity` to ensure that it will be
checked every time the app comes back from the background. Right now, after other changes, I'm not sure
anymore if this is necessary. Maybe we can put it back in `SecurityPolicyChecker` and change its code
only slightly. However, the way it is now enables me to easily test if `BiometricPrompt` works (it does!).

2. The migration to AndroidX. It consists of a lot of changes to imports, both in Java/Scala/Kotlin files
as well as in layouts. We should test it on various devices to make sure everything looks the same.

3. Glide, our image displaying library, was the only bigger problem I had during the migration. I updated it,
refactored parts of our code where we tell Glide how to display our assets (this was probably unnecessary),
and made changes to Gradle settings. It's possible that some of my changes to Gradle settings can be reverted.