Remove race conditions in SignalSpec tests
 - The `SignalSpec` tests check state immediately after performing
operations, such as checking subscriber numbers after creating a
subscription, which are performed asynchonously. This causes sporadic
test failures.
 - This can be best seen in the first test in the suite, which has a
`Thread.sleep` call which, when removed, results in the test failing
every time.
 - This commit introduces the `Eventually` construct to deal with this
in a more idiomatic manner. All test assertions are wrapped in it and
thus are retried for a short period if they fail at first.