feat: Use Wire password and a dialog instead of the Android functionality
Instead of relying on a partially broken Android functionality for asking for password when coming out of the background,
we will now use the Wire password and a dialog inside the app. The PR introduces a new property for counting failed
attempts - in personal builds it's set to 0 and not used, but in builds like EY it's used to wipe the user's data after
N failed attempts. Also, the biometric prompt can be used (if available) to speed up authentication.

I rewrote a big part of the security policy checker.