fix: Hide keyboard when using biometric (#2403)
Right now the keyboard is opened even if the user has the ability to use biometric to authenticate, and it doesn't close automatically
after a successful authentication. This PR prevents the keyboard from opening if the biometric can be used.
The keyboard still opens automatically if the biometric prompt is unavailable and the user has to type in the password.