fix: Make audio recordings explicitly M4A
The encoding of our audio recordings does not change - it's the same for mp4 and for m4a files. But using the mp4 extension
and MIME type for audio files may cause problems on the receiving side. This PR changes them explicitly to M4A in our code.

Most of the changes are in names of variables and methods. Only in a few places I changed a file extension or a mime type.