fix: A quick fix for an error in Android 10 when sending an audio recording (#2411)
While preparing an audio recording for sending, we extract some metadata out of it, i.e. duration and audio levels.
For this, we have to decode the just created m4a file and go through it, calculating the levels.
At the end of processing we push MediaCodec.BUFFER_FLAG_END_OF_STREAM to the input buffer and on Android 10 that
causes a CodecException. From what I read on SO, we do it right. So I guess this might be an Android 10 bug.

I wrapped the line causing the exception in a try/catch block. We only need it to calculate the audio levels
anyway, so even if we fail to push that constant to the buffer it seems nothing bad is going to happen.

Probably a better solution would be to rewrite the whole transcoder using a newer technology.