fix: Password prompt eforced by the configuration should still respect the timeout (#2423)
The last changes o the password prompt  broke this.
If force_app_lock is set to true, the app should respect app_lock_timeout and show the prompt
only after the given time in background passed. Recently it worked this way only if the prompt
was switched on from the options, but if it was enforced, it appeared every time, regardless
of the timeout.