Remove prefilling of Giphy search box
Clicking on the Giphy search button will not prefill the Giphy search with the message being typed. This has been raised as a security concern (accidental tapping on Giphy) and this is a simple attempt at addressing the issue. In future we could move this to a user preference to allow for both behaviours.