Refactor UserData.displayName
From time to time a bug comes up in our UI, usually about that the user name does not appear correctly
in one place or another. After investigation, we find out that we used `UserData.name` or
`UserData.displayName` instead of `UserData.getDisplayName`.
This PR makes `UserData.displayName` the only way to access the name from the outside if you want to
display it.