Preparation for "connection screen" refactoring.
The following screens are very similar to each other and to the SingleParticipantFragment screen,
but htey are all written separately and duplicate a lot of their functionality, while some of
the new functionality (e.g. conversation rules) is broken in all of them:
 - SendConnectRequestFragment (displayed for an unconnected user)
 - PendingConnectRequestFragment (after sending the connect request and waiting for the answer)
 - ConnectRequestFragment (when another user sent a connect request and we didn't answer yet)
 - BlockedUserProfileFragment (when the other user is blocked)

I want to rewrite them all, based on SingleParticipantFragment.
The first step is to make SingleParticipantFragment more generic and extendable, together with
SingleParticipantAdapter.
This PR contains those changes plus a class called SendConnectRequestFragment2 which will in future
take over from SendConnectRequestFragment, but is not used anywhere right now
(I needed it to better understand what I'm doing.)